csce482/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp>=2.6.1               # via kombu
backcall>=0.2.0           # via ipython
billiard>=3.6.3.0         # via celery
brotlipy>=0.7.0           # via -r requirements.in
celery[redis]>=4.4.7      # via -r requirements.in
certifi>=2020.6.20        # via sentry-sdk
cffi>=1.14.2              # via brotlipy
decorator>=4.4.2          # via ipython, traitlets
defusedxml>=0.6.0         # via odfpy
diff-match-patch>=20200713  # via django-import-export
dj-database-url>=0.5.0    # via -r requirements.in
django-cors-headers>=3.5.0  # via -r requirements.in
django-import-export>=2.3.0  # via -r requirements.in
django-js-reverse>=0.9.1  # via -r requirements.in
django-log-request-id>=1.6.0  # via -r requirements.in
django-model-utils>=4.0.0  # via -r requirements.in
django-webpack-loader>=0.7.0  # via -r requirements.in
django>=2.2.16            # via -r requirements.in, django-cors-headers, django-import-export, django-js-reverse, django-log-request-id, django-model-utils, djangorestframework
djangorestframework>=3.11.1  # via -r requirements.in
et-xmlfile>=1.0.1         # via openpyxl
gunicorn>=20.0.4          # via -r requirements.in
importlib-metadata>=1.7.0  # via kombu
ipython-genutils>=0.2.0   # via traitlets
ipython>=7.16.1           # via -r requirements.in
jdcal>=1.4.1              # via openpyxl
jedi>=0.17.2              # via ipython
kombu>=4.6.11             # via celery
markuppy>=1.14            # via tablib
odfpy>=1.4.1              # via tablib
openpyxl>=3.0.5           # via tablib
parso>=0.7.1              # via jedi
pexpect>=4.8.0            # via ipython
pickleshare>=0.7.5        # via ipython
prompt-toolkit>=3.0.7     # via ipython
psutil>=5.7.2             # via -r requirements.in
psycopg2>=2.8.6           # via -r requirements.in
ptyprocess>=0.6.0         # via pexpect
pycparser>=2.20           # via cffi
pygments>=2.7.0           # via ipython
python-decouple>=3.3      # via -r requirements.in
pytz>=2020.1              # via celery, django
pyyaml>=5.3.1             # via tablib
redis>=3.5.3              # via celery
sentry-sdk>=0.17.5        # via -r requirements.in
six>=1.15.0               # via traitlets
sqlparse>=0.3.1           # via django
tablib[html,ods,xls,xlsx,yaml]>=2.0.0  # via django-import-export
traitlets>=4.3.3          # via ipython
urllib3>=1.25.10          # via sentry-sdk
vine>=1.3.0               # via amqp, celery
wcwidth>=0.2.5            # via prompt-toolkit
whitenoise>=5.2.0         # via -r requirements.in
xlrd>=1.2.0               # via tablib
xlwt>=1.3.0               # via tablib
zipp>=3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
